+++ tests/conformance/properties.f
@@ -118,7 +118,7 @@ arr[text] func styleDigestFields(s:Style) {
         `${shadowKey(s.shadows)}`,
         `${s.counterReset}`, `${s.counterIncrement}`, `${s.counterSet}`, `${s.quotes}`,
         `${s.colorSchemeDark}`, `${s.containerType}`, `${s.containerName}`,
-        `${initialLetterPacked(s)}`, `${baselineSourceOf(s)}`, `${resizeOf(s)}`]
+        `${initialLetterPacked(s)}`, `${baselineSourceOf(s)}`, `${resizeOf(s)}`, `${textWrapStyleOf(s)}`]
 }
 
 text func motionKeyPath(mi:MotionInfo) {
@@ -208,7 +208,7 @@ arr[text] func styleDigestFieldNames() {
         'backgroundOrigin', 'objectFit', 'objectPosX', 'objectPosY', 'objectViewBox', 'shadows',
         'counterReset', 'counterIncrement', 'counterSet', 'quotes', 'colorSchemeDark',
         'containerType', 'containerName', 'initialLetterPacked', 'baselineSource',
-        'resize']
+        'resize', 'textWrapStyle']
 }
 
 text func trackKey(list:arr[Track]) {